requirements.txt
matplotlib
numpy
pandas
pandas-datareader
pandas_market_calendars
pandas_ta
plotly
quantstats
requests_cache
sklearn
sqlalchemy
ta
tensorflow
yfinance
